refactor: usuario

+++ src/main/java/com/tcc/backend/usuario/UsuarioService.java
@@ -1,13 +1,12 @@
 package com.tcc.backend.usuario;
 
-import java.lang.reflect.Type;
 import java.util.List;
-import java.util.Optional;
 
-import org.modelmapper.ModelMapper;
-import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.tcc.backend.exceptions.ResourceAlreadyExists;
+import com.tcc.backend.exceptions.ResourceNotFoundException;
 
 @Service
 public class UsuarioService {
@@ -15,44 +14,40 @@ public class UsuarioService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
-    @Autowired
-    private ModelMapper modelMapper;
-
-    public List<UsuarioDto> listarTodos() {
-        List<Usuario> usuariosEntidade = usuarioRepository.findAll();
-        Type listType = new TypeToken<List<UsuarioDto>>(){}.getType();
-
-        List<UsuarioDto> usuariosDto = modelMapper.map(usuariosEntidade, listType);
-        return usuariosDto;
+    public List<Usuario> listarTodos() {
+        return usuarioRepository.findAll();
     }
 
-    public UsuarioDto criarUsuario(UsuarioDto usuarioDto){
-        Usuario usuario = usuarioRepository.save(modelMapper.map(usuarioDto, Usuario.class));
+    public Usuario criarUsuario(Usuario usuario){
+        List<Usuario> listaUsuarios = usuarioRepository.findByNomeIgnoreCase(usuario.getNome());
 
-        UsuarioDto usuarioResponse = modelMapper.map(usuario, UsuarioDto.class);
-
-        return usuarioResponse;
-    }
-
-    public UsuarioDto consultarUsuarioPorId(Long id){
-        Optional<Usuario> usuario = usuarioRepository.findById(id);
-
-        if(usuario.isPresent()){
-            return modelMapper.map(usuario.get(), UsuarioDto.class);
-        } else{
-            return null;
+        if (!listaUsuarios.isEmpty()) {
+            throw new ResourceAlreadyExists("Usuario", "nome", usuario.getNome());
         }
+        
+        return usuarioRepository.save(usuario);
     }
 
-    public UsuarioDto atualizarUsuario(Long id, UsuarioDto usuarioDto){
-        Usuario usuario = usuarioRepository.findById(id).orElseThrow();
-        usuario.setIsAdmin(usuarioDto.getIsAdmin());
-        usuario.setNome(usuarioDto.getNome());
-        usuarioRepository.save(usuario);
-        return modelMapper.map(usuario, UsuarioDto.class);
+    public Usuario consultarUsuarioPorId(Long id){
+        return usuarioRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Usuario", id));
+    }
+
+    public Usuario atualizarUsuario(Long id, Usuario usuarioReq){
+        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Usuario", id));
+        
+        if (!usuarioRepository.findByNomeIgnoreCase(usuarioReq.getNome()).isEmpty()) {
+            throw new ResourceAlreadyExists("Usuario", "nome", usuarioReq.getNome());
+        }
+        usuario.setIsAdmin(usuarioReq.getIsAdmin());
+        usuario.setNome(usuarioReq.getNome());
+        return usuarioRepository.save(usuario);
     }
     
     public void deletarUsuario(Long id){
+        if (usuarioRepository.findById(id).isEmpty()){
+            throw new ResourceNotFoundException("Usuario", id);
+        }
         usuarioRepository.deleteById(id);
     }
+
 }
